refactor: added dto validations, @Autowired for consistency, update tested null value
