+++ src/main/java/com/DevSprint/voluntrix_backend/controllers/EventController.java
@@ -3,6 +3,7 @@ package com.DevSprint.voluntrix_backend.controllers;
 import java.time.LocalDate;
 import java.util.List;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -30,10 +31,14 @@ import lombok.RequiredArgsConstructor;
 @RestController
 @RequestMapping("/api/public/events")
 @SecurityRequirement(name = "bearerAuth")
-@RequiredArgsConstructor
 public class EventController {
 
     private final EventService eventService;
+
+    @Autowired
+    public EventController(EventService eventService) {
+        this.eventService = eventService;
+    }
 
     @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Void> addEvent(@RequestBody EventDTO eventDTO) {
