+++ src/main/java/com/DevSprint/voluntrix_backend/services/EmailService.java
@@ -2,8 +2,8 @@ package com.DevSprint.voluntrix_backend.services;
 
 import jakarta.mail.MessagingException;
 import jakarta.mail.internet.MimeMessage;
-import lombok.RequiredArgsConstructor;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.mail.MailException;
 import org.springframework.mail.javamail.JavaMailSender;
@@ -15,10 +15,14 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 
 @Service
-@RequiredArgsConstructor
 public class EmailService {
 
     private final JavaMailSender mailSender;
+
+    @Autowired
+    public EmailService(JavaMailSender mailSender) {
+        this.mailSender = mailSender;
+    }
 
     public void sendEmailService(String to, String name, String orderId, Double amount) throws MailException,MessagingException,IOException{
         String subject = "Thank You for Your Donation to Voluntrix!";
@@ -28,7 +32,7 @@ public class EmailService {
         MimeMessageHelper helper = new MimeMessageHelper(message, true);
         helper.setTo(to);
         helper.setSubject(subject);
-        helper.setText(null, true); 
+        helper.setText(content, true); 
 
         mailSender.send(message);
     }
